[X86] Disassemble cap-aware addresses for memory operands.
This expanded the decision tables beyond 2^16 entries, so the InstrUID
type is now a uint32_t instead of a uint16_t.

Some of the clauses in inheritsFrom() probably still need more work,
but this gets some simple cases working.

+++ llvm/include/llvm/Support/X86DisassemblerDecoderCommon.h
@@ -332,7 +332,7 @@ enum OpcodeType {
 //
 // If a ModR/M byte is not required, "required" is left unset, and the values
 // for each instructionID are identical.
-typedef uint16_t InstrUID;
+typedef uint32_t InstrUID;
 
 // ModRMDecisionType - describes the type of ModR/M decision, allowing the
 // consumer to determine the number of entries in it.
